mms/backend/requirements.txt
aniso8601==9.0.1
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
ddtrace==0.60.0
Flask==2.0.3
flask-restplus==0.13.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
jsonschema==4.4.0
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.0.1
protobuf==3.19.4
pyparsing==3.0.7
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.32
tenacity==8.0.1
urllib3==1.26.9
Werkzeug==2.0.3
